fix: landing matrix float error

+++ src/duckdb-rs/src/refresher.rs
@@ -31,7 +31,7 @@ const LANDING_SCHEMA: &str = "CREATE TABLE
         fiskeridir_vessel_id INT,
         gear_group_id INT NOT NULL,
         species_group_id INT NOT NULL,
-        living_weight BIGINT NOT NULL,
+        living_weight DOUBLE NOT NULL,
         PRIMARY KEY (landing_id, species_group_id)
     )";
